4.34. Finalizando a atualização parcial com a API de Reflections do Spring
Endpoints de exclusao para os Controllers
Criado um endpoint com o verbo http Patch. Usamos ObjectMapper pr mapear
um json no Objeto Restaurante e reflection para copiar os atributos do
map para o objeto

+++ algafood-api/src/main/java/com/algaworks/algafood/infrastructure/repository/EstadoRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -36,6 +37,10 @@ public class EstadoRepositoryImpl implements EstadoRepository {
 	@Override
 	@Transactional
 	public void remover(Long estadoId) {
+		Estado estado = buscarPorId(estadoId);
+		if (estado == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
 		entityManager.remove(estadoId);
 	}
 }
